Fixing clang-tidy-16 Warnings

See merge request allpix-squared/allpix-squared!1091

+++ src/modules/CapacitiveTransfer/CapacitiveTransferModule.cpp
@@ -156,8 +156,8 @@ void CapacitiveTransferModule::initialize() {
         if(root_file->IsZombie()) {
             throw InvalidValueError(config_, "coupling_scan_file", "ROOT file is corrupted. Please, check it");
         }
-        for(int i = 1; i < 10; i++) {
-            capacitances_[i - 1] = static_cast<TGraph*>(root_file->Get(Form("Pixel_%i", i)));
+        for(size_t i = 1; i < 10; i++) {
+            capacitances_[i - 1] = static_cast<TGraph*>(root_file->Get(Form("Pixel_%zu", i)));
             if(capacitances_[i - 1]->IsZombie()) {
                 throw InvalidValueError(
                     config_, "coupling_scan_file", "ROOT TGraphs couldn't be imported. Please, check it");
@@ -373,8 +373,8 @@ void CapacitiveTransferModule::finalize() {
             gap_map->Write();
             capacitance_map->Write();
             relative_capacitance_map->Write();
-            for(int i = 1; i < 10; i++) {
-                capacitances_[i - 1]->Write(Form("Pixel_%i", i));
+            for(size_t i = 1; i < 10; i++) {
+                capacitances_[i - 1]->Write(Form("Pixel_%zu", i));
             }
         } else {
             coupling_map->Write();
